feat(ItemsTable): create child component ItemsTable
