+++ src/Ml.Cli.WebApp/ClientApp/src/EmptyArrayManager/EmptyArrayManager.scss
@@ -1,5 +1,6 @@
 ﻿.empty-array{
   &__error{
     text-align: center;
+    display: block;
   }
 }
