+++ src/Ml.Cli.WebApp/ClientApp/src/EmptyArrayManager/index.js
@@ -10,7 +10,7 @@ class EmptyArrayManager extends Component {
                 return <div>{child}</div>
             }
         }
-        return null;
+        return <span className="empty-array__error"><b>{this.props.onArrayEmpty}</b></span>
     };
 }
